Crate rusoto_swf

Amazon Simple Workflow Service
The Amazon Simple Workflow Service (Amazon SWF) makes it easy to build applications that use Amazon's cloud to coordinate work across distributed components. In Amazon SWF, a task represents a logical unit of work that is performed by a component of your workflow. Coordinating tasks in a workflow involves managing intertask dependencies, scheduling, and concurrency in accordance with the logical flow of the application.

Amazon SWF gives you full control over implementing tasks and coordinating them without worrying about underlying complexities such as tracking their progress and maintaining their state.

This documentation serves as reference only. For a broader overview of the Amazon SWF programming model, see the Amazon SWF Developer Guide .

If you're using the service, you're probably looking for SwfClient and Swf.

Structs

ActivityTask
Unit of work sent to an activity worker.

ActivityTaskCancelRequestedEventAttributes
Provides the details of the ActivityTaskCancelRequested event.

ActivityTaskCanceledEventAttributes
Provides the details of the ActivityTaskCanceled event.

ActivityTaskCompletedEventAttributes
Provides the details of the ActivityTaskCompleted event.

ActivityTaskFailedEventAttributes
Provides the details of the ActivityTaskFailed event.

ActivityTaskScheduledEventAttributes
Provides the details of the ActivityTaskScheduled event.

ActivityTaskStartedEventAttributes
Provides the details of the ActivityTaskStarted event.

ActivityTaskStatus
Status information about an activity task.

ActivityTaskTimedOutEventAttributes
Provides the details of the ActivityTaskTimedOut event.

ActivityType
Represents an activity type.

ActivityTypeConfiguration
Configuration settings registered with the activity type.

ActivityTypeDetail
Detailed information about an activity type.

ActivityTypeInfo
Detailed information about an activity type.

ActivityTypeInfos
Contains a paginated list of activity type information structures.

CancelTimerDecisionAttributes
Provides the details of the CancelTimer decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

CancelTimerFailedEventAttributes
Provides the details of the CancelTimerFailed event.

CancelWorkflowExecutionDecisionAttributes
Provides the details of the CancelWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

CancelWorkflowExecutionFailedEventAttributes
Provides the details of the CancelWorkflowExecutionFailed event.

ChildWorkflowExecutionCanceledEventAttributes
Provide details of the ChildWorkflowExecutionCanceled event.

ChildWorkflowExecutionCompletedEventAttributes
Provides the details of the ChildWorkflowExecutionCompleted event.

ChildWorkflowExecutionFailedEventAttributes
Provides the details of the ChildWorkflowExecutionFailed event.

ChildWorkflowExecutionStartedEventAttributes
Provides the details of the ChildWorkflowExecutionStarted event.

ChildWorkflowExecutionTerminatedEventAttributes
Provides the details of the ChildWorkflowExecutionTerminated event.

ChildWorkflowExecutionTimedOutEventAttributes
Provides the details of the ChildWorkflowExecutionTimedOut event.

CloseStatusFilter
Used to filter the closed workflow executions in visibility APIs by their close status.

CompleteWorkflowExecutionDecisionAttributes
Provides the details of the CompleteWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

CompleteWorkflowExecutionFailedEventAttributes
Provides the details of the CompleteWorkflowExecutionFailed event.

ContinueAsNewWorkflowExecutionDecisionAttributes
Provides the details of the ContinueAsNewWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tag – A tag used to identify the workflow execution

taskList – String constraint. The key is swf:taskList.name.

workflowType.version – String constraint. The key is swf:workflowType.version.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

ContinueAsNewWorkflowExecutionFailedEventAttributes
Provides the details of the ContinueAsNewWorkflowExecutionFailed event.

CountClosedWorkflowExecutionsInput

CountOpenWorkflowExecutionsInput

CountPendingActivityTasksInput

CountPendingDecisionTasksInput

Decision

Specifies a decision made by the decider. A decision can be one of these types:

CancelTimer – Cancels a previously started timer and records a TimerCanceled event in the history.

CancelWorkflowExecution – Closes the workflow execution and records a WorkflowExecutionCanceled event in the history.

CompleteWorkflowExecution – Closes the workflow execution and records a WorkflowExecutionCompleted event in the history .

ContinueAsNewWorkflowExecution – Closes the workflow execution and starts a new workflow execution of the same type using the same workflow ID and a unique run Id. A WorkflowExecutionContinuedAsNew event is recorded in the history.

FailWorkflowExecution – Closes the workflow execution and records a WorkflowExecutionFailed event in the history.

RecordMarker – Records a MarkerRecorded event in the history. Markers can be used for adding custom information in the history for instance to let deciders know that they don't need to look at the history beyond the marker event.

RequestCancelActivityTask – Attempts to cancel a previously scheduled activity task. If the activity task was scheduled but has not been assigned to a worker, then it is canceled. If the activity task was already assigned to a worker, then the worker is informed that cancellation has been requested in the response to RecordActivityTaskHeartbeat.

RequestCancelExternalWorkflowExecution – Requests that a request be made to cancel the specified external workflow execution and records a RequestCancelExternalWorkflowExecutionInitiated event in the history.

ScheduleActivityTask – Schedules an activity task.

SignalExternalWorkflowExecution – Requests a signal to be delivered to the specified external workflow execution and records a SignalExternalWorkflowExecutionInitiated event in the history.

StartChildWorkflowExecution – Requests that a child workflow execution be started and records a StartChildWorkflowExecutionInitiated event in the history. The child workflow execution is a separate workflow execution with its own history.

StartTimer – Starts a timer for this workflow execution and records a TimerStarted event in the history. This timer fires after the specified delay and record a TimerFired event.

Access Control

If you grant permission to use RespondDecisionTaskCompleted, you can use IAM policies to express permissions for the list of decisions returned by this action as if they were members of the API. Treating decisions as a pseudo API maintains a uniform conceptual model and helps keep policies readable. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

Decision Failure

Decisions can fail for several reasons

The ordering of decisions should follow a logical flow. Some decisions might not make sense in the current context of the workflow execution and therefore fails.

A limit on your account was reached.

The decision lacks sufficient permissions.

One of the following events might be added to the history to indicate an error. The event attribute's cause parameter indicates the cause. If cause is set to OPERATIONNOTPERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

ScheduleActivityTaskFailed – A ScheduleActivityTask decision failed. This could happen if the activity type specified in the decision isn't registered, is in a deprecated state, or the decision isn't properly configured.

RequestCancelActivityTaskFailed – A RequestCancelActivityTask decision failed. This could happen if there is no open activity task with the specified activityId.

StartTimerFailed – A StartTimer decision failed. This could happen if there is another open timer with the same timerId.

CancelTimerFailed – A CancelTimer decision failed. This could happen if there is no open timer with the specified timerId.

StartChildWorkflowExecutionFailed – A StartChildWorkflowExecution decision failed. This could happen if the workflow type specified isn't registered, is deprecated, or the decision isn't properly configured.

SignalExternalWorkflowExecutionFailed – A SignalExternalWorkflowExecution decision failed. This could happen if the workflowID specified in the decision was incorrect.

RequestCancelExternalWorkflowExecutionFailed – A RequestCancelExternalWorkflowExecution decision failed. This could happen if the workflowID specified in the decision was incorrect.

CancelWorkflowExecutionFailed – A CancelWorkflowExecution decision failed. This could happen if there is an unhandled decision task pending in the workflow execution.

CompleteWorkflowExecutionFailed – A CompleteWorkflowExecution decision failed. This could happen if there is an unhandled decision task pending in the workflow execution.

ContinueAsNewWorkflowExecutionFailed – A ContinueAsNewWorkflowExecution decision failed. This could happen if there is an unhandled decision task pending in the workflow execution or the ContinueAsNewWorkflowExecution decision was not configured correctly.

FailWorkflowExecutionFailed – A FailWorkflowExecution decision failed. This could happen if there is an unhandled decision task pending in the workflow execution.

The preceding error events might occur due to an error in the decider logic, which might put the workflow execution in an unstable state The cause field in the event structure for the error event indicates the cause of the error.

A workflow execution may be closed by the decider by returning one of the following decisions when completing a decision task: CompleteWorkflowExecution, FailWorkflowExecution, CancelWorkflowExecution and ContinueAsNewWorkflowExecution. An UnhandledDecision fault is returned if a workflow closing decision is specified and a signal or activity event had been added to the history while the decision task was being performed by the decider. Unlike the above situations which are logic issues, this fault is always possible because of race conditions in a distributed system. The right action here is to call RespondDecisionTaskCompleted without any decisions. This would result in another decision task with these new events included in the history. The decider should handle the new events and may decide to close the workflow execution.

How to Code a Decision

You code a decision by first setting the decision type field to one of the above decision values, and then set the corresponding attributes field shown below:

ScheduleActivityTaskDecisionAttributes

RequestCancelActivityTaskDecisionAttributes

CompleteWorkflowExecutionDecisionAttributes

FailWorkflowExecutionDecisionAttributes

CancelWorkflowExecutionDecisionAttributes

ContinueAsNewWorkflowExecutionDecisionAttributes

RecordMarkerDecisionAttributes

StartTimerDecisionAttributes

CancelTimerDecisionAttributes

SignalExternalWorkflowExecutionDecisionAttributes

RequestCancelExternalWorkflowExecutionDecisionAttributes

StartChildWorkflowExecutionDecisionAttributes

DecisionTask
A structure that represents a decision task. Decision tasks are sent to deciders in order for them to make decisions.

DecisionTaskCompletedEventAttributes
Provides the details of the DecisionTaskCompleted event.

DecisionTaskScheduledEventAttributes
Provides details about the DecisionTaskScheduled event.

DecisionTaskStartedEventAttributes
Provides the details of the DecisionTaskStarted event.

DecisionTaskTimedOutEventAttributes
Provides the details of the DecisionTaskTimedOut event.

DeprecateActivityTypeInput

DeprecateDomainInput

DeprecateWorkflowTypeInput

DescribeActivityTypeInput

DescribeDomainInput

DescribeWorkflowExecutionInput

DescribeWorkflowTypeInput

DomainConfiguration
Contains the configuration settings of a domain.

DomainDetail
Contains details of a domain.

DomainInfo
Contains general information about a domain.

DomainInfos
Contains a paginated collection of DomainInfo structures.

ExecutionTimeFilter
Used to filter the workflow executions in visibility APIs by various time-based rules. Each parameter, if specified, defines a rule that must be satisfied by each returned query result. The parameter values are in the Unix Time format. For example: "oldestDate": 1325376070.

ExternalWorkflowExecutionCancelRequestedEventAttributes
Provides the details of the ExternalWorkflowExecutionCancelRequested event.

ExternalWorkflowExecutionSignaledEventAttributes
Provides the details of the ExternalWorkflowExecutionSignaled event.

FailWorkflowExecutionDecisionAttributes
Provides the details of the FailWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

FailWorkflowExecutionFailedEventAttributes
Provides the details of the FailWorkflowExecutionFailed event.

GetWorkflowExecutionHistoryInput

History
Paginated representation of a workflow history for a workflow execution. This is the up to date, complete and authoritative record of the events related to all tasks and events in the life of the workflow execution.

HistoryEvent

Event within a workflow execution. A history event can be one of these types:

ActivityTaskCancelRequested – A RequestCancelActivityTask decision was received by the system.

ActivityTaskCanceled – The activity task was successfully canceled.

ActivityTaskCompleted – An activity worker successfully completed an activity task by calling RespondActivityTaskCompleted.

ActivityTaskFailed – An activity worker failed an activity task by calling RespondActivityTaskFailed.

ActivityTaskScheduled – An activity task was scheduled for execution.

ActivityTaskStarted – The scheduled activity task was dispatched to a worker.

ActivityTaskTimedOut – The activity task timed out.

CancelTimerFailed – Failed to process CancelTimer decision. This happens when the decision isn't configured properly, for example no timer exists with the specified timer Id.

CancelWorkflowExecutionFailed – A request to cancel a workflow execution failed.

ChildWorkflowExecutionCanceled – A child workflow execution, started by this workflow execution, was canceled and closed.

ChildWorkflowExecutionCompleted – A child workflow execution, started by this workflow execution, completed successfully and was closed.

ChildWorkflowExecutionFailed – A child workflow execution, started by this workflow execution, failed to complete successfully and was closed.

ChildWorkflowExecutionStarted – A child workflow execution was successfully started.

ChildWorkflowExecutionTerminated – A child workflow execution, started by this workflow execution, was terminated.

ChildWorkflowExecutionTimedOut – A child workflow execution, started by this workflow execution, timed out and was closed.

CompleteWorkflowExecutionFailed – The workflow execution failed to complete.

ContinueAsNewWorkflowExecutionFailed – The workflow execution failed to complete after being continued as a new workflow execution.

DecisionTaskCompleted – The decider successfully completed a decision task by calling RespondDecisionTaskCompleted.

DecisionTaskScheduled – A decision task was scheduled for the workflow execution.

DecisionTaskStarted – The decision task was dispatched to a decider.

DecisionTaskTimedOut – The decision task timed out.

ExternalWorkflowExecutionCancelRequested – Request to cancel an external workflow execution was successfully delivered to the target execution.

ExternalWorkflowExecutionSignaled – A signal, requested by this workflow execution, was successfully delivered to the target external workflow execution.

FailWorkflowExecutionFailed – A request to mark a workflow execution as failed, itself failed.

MarkerRecorded – A marker was recorded in the workflow history as the result of a RecordMarker decision.

RecordMarkerFailed – A RecordMarker decision was returned as failed.

RequestCancelActivityTaskFailed – Failed to process RequestCancelActivityTask decision. This happens when the decision isn't configured properly.

RequestCancelExternalWorkflowExecutionFailed – Request to cancel an external workflow execution failed.

RequestCancelExternalWorkflowExecutionInitiated – A request was made to request the cancellation of an external workflow execution.

ScheduleActivityTaskFailed – Failed to process ScheduleActivityTask decision. This happens when the decision isn't configured properly, for example the activity type specified isn't registered.

SignalExternalWorkflowExecutionFailed – The request to signal an external workflow execution failed.

SignalExternalWorkflowExecutionInitiated – A request to signal an external workflow was made.

StartActivityTaskFailed – A scheduled activity task failed to start.

StartChildWorkflowExecutionFailed – Failed to process StartChildWorkflowExecution decision. This happens when the decision isn't configured properly, for example the workflow type specified isn't registered.

StartChildWorkflowExecutionInitiated – A request was made to start a child workflow execution.

StartTimerFailed – Failed to process StartTimer decision. This happens when the decision isn't configured properly, for example a timer already exists with the specified timer Id.

TimerCanceled – A timer, previously started for this workflow execution, was successfully canceled.

TimerFired – A timer, previously started for this workflow execution, fired.

TimerStarted – A timer was started for the workflow execution due to a StartTimer decision.

WorkflowExecutionCancelRequested – A request to cancel this workflow execution was made.

WorkflowExecutionCanceled – The workflow execution was successfully canceled and closed.

WorkflowExecutionCompleted – The workflow execution was closed due to successful completion.

WorkflowExecutionContinuedAsNew – The workflow execution was closed and a new execution of the same type was created with the same workflowId.

WorkflowExecutionFailed – The workflow execution closed due to a failure.

WorkflowExecutionSignaled – An external signal was received for the workflow execution.

WorkflowExecutionStarted – The workflow execution was started.

WorkflowExecutionTerminated – The workflow execution was terminated.

WorkflowExecutionTimedOut – The workflow execution was closed because a time out was exceeded.

LambdaFunctionCompletedEventAttributes
Provides the details of the LambdaFunctionCompleted event. It isn't set for other event types.

LambdaFunctionFailedEventAttributes
Provides the details of the LambdaFunctionFailed event. It isn't set for other event types.

LambdaFunctionScheduledEventAttributes
Provides the details of the LambdaFunctionScheduled event. It isn't set for other event types.

LambdaFunctionStartedEventAttributes
Provides the details of the LambdaFunctionStarted event. It isn't set for other event types.

LambdaFunctionTimedOutEventAttributes
Provides details of the LambdaFunctionTimedOut event.

ListActivityTypesInput

ListClosedWorkflowExecutionsInput

ListDomainsInput

ListOpenWorkflowExecutionsInput

ListWorkflowTypesInput

MarkerRecordedEventAttributes
Provides the details of the MarkerRecorded event.

PendingTaskCount
Contains the count of tasks in a task list.

PollForActivityTaskInput

PollForDecisionTaskInput

RecordActivityTaskHeartbeatInput

RecordMarkerDecisionAttributes
Provides the details of the RecordMarker decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

RecordMarkerFailedEventAttributes
Provides the details of the RecordMarkerFailed event.

RegisterActivityTypeInput

RegisterDomainInput

RegisterWorkflowTypeInput

RequestCancelActivityTaskDecisionAttributes
Provides the details of the RequestCancelActivityTask decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

RequestCancelActivityTaskFailedEventAttributes
Provides the details of the RequestCancelActivityTaskFailed event.

RequestCancelExternalWorkflowExecutionDecisionAttributes
Provides the details of the RequestCancelExternalWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

RequestCancelExternalWorkflowExecutionFailedEventAttributes
Provides the details of the RequestCancelExternalWorkflowExecutionFailed event.

RequestCancelExternalWorkflowExecutionInitiatedEventAttributes
Provides the details of the RequestCancelExternalWorkflowExecutionInitiated event.

RequestCancelWorkflowExecutionInput

RespondActivityTaskCanceledInput

RespondActivityTaskCompletedInput

RespondActivityTaskFailedInput

RespondDecisionTaskCompletedInput
Input data for a TaskCompleted response to a decision task.

Run
Specifies the runId of a workflow execution.

ScheduleActivityTaskDecisionAttributes
Provides the details of the ScheduleActivityTask decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

activityType.name – String constraint. The key is swf:activityType.name.

activityType.version – String constraint. The key is swf:activityType.version.

taskList – String constraint. The key is swf:taskList.name.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

ScheduleActivityTaskFailedEventAttributes
Provides the details of the ScheduleActivityTaskFailed event.

ScheduleLambdaFunctionDecisionAttributes
Decision attributes specified in scheduleLambdaFunctionDecisionAttributes within the list of decisions decisions passed to RespondDecisionTaskCompleted.

ScheduleLambdaFunctionFailedEventAttributes
Provides the details of the ScheduleLambdaFunctionFailed event. It isn't set for other event types.

SignalExternalWorkflowExecutionDecisionAttributes
Provides the details of the SignalExternalWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

SignalExternalWorkflowExecutionFailedEventAttributes
Provides the details of the SignalExternalWorkflowExecutionFailed event.

SignalExternalWorkflowExecutionInitiatedEventAttributes
Provides the details of the SignalExternalWorkflowExecutionInitiated event.

SignalWorkflowExecutionInput

StartChildWorkflowExecutionDecisionAttributes
Provides the details of the StartChildWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tagList.member.N – The key is "swf:tagList.N" where N is the tag number from 0 to 4, inclusive.

taskList – String constraint. The key is swf:taskList.name.

workflowType.name – String constraint. The key is swf:workflowType.name.

workflowType.version – String constraint. The key is swf:workflowType.version.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

StartChildWorkflowExecutionFailedEventAttributes
Provides the details of the StartChildWorkflowExecutionFailed event.

StartChildWorkflowExecutionInitiatedEventAttributes
Provides the details of the StartChildWorkflowExecutionInitiated event.

StartLambdaFunctionFailedEventAttributes
Provides the details of the StartLambdaFunctionFailed event. It isn't set for other event types.

StartTimerDecisionAttributes
Provides the details of the StartTimer decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

StartTimerFailedEventAttributes
Provides the details of the StartTimerFailed event.

StartWorkflowExecutionInput

SwfClient
A client for the Amazon SWF API.

TagFilter
Used to filter the workflow executions in visibility APIs based on a tag.

TaskList
Represents a task list.

TerminateWorkflowExecutionInput

TimerCanceledEventAttributes
Provides the details of the TimerCanceled event.

TimerFiredEventAttributes
Provides the details of the TimerFired event.

TimerStartedEventAttributes
Provides the details of the TimerStarted event.

WorkflowExecution
Represents a workflow execution.

WorkflowExecutionCancelRequestedEventAttributes
Provides the details of the WorkflowExecutionCancelRequested event.

WorkflowExecutionCanceledEventAttributes
Provides the details of the WorkflowExecutionCanceled event.

WorkflowExecutionCompletedEventAttributes
Provides the details of the WorkflowExecutionCompleted event.

WorkflowExecutionConfiguration
The configuration settings for a workflow execution including timeout values, tasklist etc. These configuration settings are determined from the defaults specified when registering the workflow type and those specified when starting the workflow execution.

WorkflowExecutionContinuedAsNewEventAttributes
Provides the details of the WorkflowExecutionContinuedAsNew event.

WorkflowExecutionCount
Contains the count of workflow executions returned from CountOpenWorkflowExecutions or CountClosedWorkflowExecutions

WorkflowExecutionDetail
Contains details about a workflow execution.

WorkflowExecutionFailedEventAttributes
Provides the details of the WorkflowExecutionFailed event.

WorkflowExecutionFilter
Used to filter the workflow executions in visibility APIs by their workflowId.

WorkflowExecutionInfo
Contains information about a workflow execution.

WorkflowExecutionInfos
Contains a paginated list of information about workflow executions.

WorkflowExecutionOpenCounts
Contains the counts of open tasks, child workflow executions and timers for a workflow execution.

WorkflowExecutionSignaledEventAttributes
Provides the details of the WorkflowExecutionSignaled event.

WorkflowExecutionStartedEventAttributes
Provides details of WorkflowExecutionStarted event.

WorkflowExecutionTerminatedEventAttributes
Provides the details of the WorkflowExecutionTerminated event.

WorkflowExecutionTimedOutEventAttributes
Provides the details of the WorkflowExecutionTimedOut event.

WorkflowType
Represents a workflow type.

WorkflowTypeConfiguration
The configuration settings of a workflow type.

WorkflowTypeDetail
Contains details about a workflow type.

WorkflowTypeFilter
Used to filter workflow execution query results by type. Each parameter, if specified, defines a rule that must be satisfied by each returned result.

WorkflowTypeInfo
Contains information about a workflow type.

WorkflowTypeInfos
Contains a paginated list of information structures about workflow types.

Enums

CountClosedWorkflowExecutionsError
Errors returned by CountClosedWorkflowExecutions

CountOpenWorkflowExecutionsError
Errors returned by CountOpenWorkflowExecutions

CountPendingActivityTasksError
Errors returned by CountPendingActivityTasks

CountPendingDecisionTasksError
Errors returned by CountPendingDecisionTasks

DeprecateActivityTypeError
Errors returned by DeprecateActivityType

DeprecateDomainError
Errors returned by DeprecateDomain

DeprecateWorkflowTypeError
Errors returned by DeprecateWorkflowType

DescribeActivityTypeError
Errors returned by DescribeActivityType

DescribeDomainError
Errors returned by DescribeDomain

DescribeWorkflowExecutionError
Errors returned by DescribeWorkflowExecution

DescribeWorkflowTypeError
Errors returned by DescribeWorkflowType

GetWorkflowExecutionHistoryError
Errors returned by GetWorkflowExecutionHistory

ListActivityTypesError
Errors returned by ListActivityTypes

ListClosedWorkflowExecutionsError
Errors returned by ListClosedWorkflowExecutions

ListDomainsError
Errors returned by ListDomains

ListOpenWorkflowExecutionsError
Errors returned by ListOpenWorkflowExecutions

ListWorkflowTypesError
Errors returned by ListWorkflowTypes

PollForActivityTaskError
Errors returned by PollForActivityTask

PollForDecisionTaskError
Errors returned by PollForDecisionTask

RecordActivityTaskHeartbeatError
Errors returned by RecordActivityTaskHeartbeat

RegisterActivityTypeError
Errors returned by RegisterActivityType

RegisterDomainError
Errors returned by RegisterDomain

RegisterWorkflowTypeError
Errors returned by RegisterWorkflowType

RequestCancelWorkflowExecutionError
Errors returned by RequestCancelWorkflowExecution

RespondActivityTaskCanceledError
Errors returned by RespondActivityTaskCanceled

RespondActivityTaskCompletedError
Errors returned by RespondActivityTaskCompleted

RespondActivityTaskFailedError
Errors returned by RespondActivityTaskFailed

RespondDecisionTaskCompletedError
Errors returned by RespondDecisionTaskCompleted

SignalWorkflowExecutionError
Errors returned by SignalWorkflowExecution

StartWorkflowExecutionError
Errors returned by StartWorkflowExecution

TerminateWorkflowExecutionError
Errors returned by TerminateWorkflowExecution

Traits

Swf
Trait representing the capabilities of the Amazon SWF API. Amazon SWF clients implement this trait.