Struct rusoto_swf::ActivityTask

pub struct ActivityTask {
    pub activity_id: String,
    pub activity_type: ActivityType,
    pub input: Option<String>,
    pub started_event_id: i64,
    pub task_token: String,
    pub workflow_execution: WorkflowExecution,
}

Unit of work sent to an activity worker.

Fields
activity_id: String

The unique ID of the task.

activity_type: ActivityType

The type of this activity task.

input: Option<String>

The inputs provided when the activity task was scheduled. The form of the input is user defined and should be meaningful to the activity implementation.

started_event_id: i64

The ID of the ActivityTaskStarted event recorded in the history.

task_token: String

The opaque string used as a handle on the task. This token is used by workers to communicate progress and response information back to the system about the task.

workflow_execution: WorkflowExecution

The workflow execution that started this activity task.