Struct rusoto_swf::ContinueAsNewWorkflowExecutionDecisionAttributes

pub struct ContinueAsNewWorkflowExecutionDecisionAttributes {
    pub child_policy: Option<String>,
    pub execution_start_to_close_timeout: Option<String>,
    pub input: Option<String>,
    pub lambda_role: Option<String>,
    pub tag_list: Option<Vec<String>>,
    pub task_list: Option<TaskList>,
    pub task_priority: Option<String>,
    pub task_start_to_close_timeout: Option<String>,
    pub workflow_type_version: Option<String>,
}

Provides details of the ContinueAsNewWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tag: Optional.. A tag used to identify the workflow execution

taskList: String constraint. The key is swf:taskList.name.

workflowType.version: String constraint. The key is swf:workflowType.version.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

Fields
child_policy: Option<String>

If set, specifies the policy to use for the child workflow executions of the new execution if it is terminated by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This policy overrides the default child policy specified when registering the workflow type using RegisterWorkflowType.

The supported child policies are:

TERMINATE: the child executions will be terminated.

REQUEST_CANCEL: a request to cancel will be attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON: no action will be taken. The child executions will continue to run.

A child policy for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default child policy was specified at registration time then a fault will be returned.

execution_start_to_close_timeout: Option<String>

If set, specifies the total duration for this workflow execution. This overrides the defaultExecutionStartToCloseTimeout specified when registering the workflow type.

The duration is specified in seconds; an integer greater than or equal to 0. The value "NONE" can be used to specify unlimited duration.
An execution start-to-close timeout for this workflow execution must be specified either as a default for the workflow type or through this field. If neither this field is set nor a default execution start-to-close timeout was specified at registration time then a fault will be returned.

input: Option<String>

The input provided to the new workflow execution.

lambda_role: Option<String>

The ARN of an IAM role that authorizes Amazon SWF to invoke AWS Lambda functions.
In order for this workflow execution to invoke AWS Lambda functions, an appropriate IAM role must be specified either as a default for the workflow type or through this field.

tag_list: Option<Vec<String>>

The list of tags to associate with the new workflow execution. A maximum of 5 tags can be specified. You can list workflow executions with a specific tag by calling ListOpenWorkflowExecutions or ListClosedWorkflowExecutions and specifying a TagFilter.

task_list: Option<TaskList> task_priority: Option<String>

Optional. The task priority that, if set, specifies the priority for the decision tasks for this workflow execution. This overrides the defaultTaskPriority specified when registering the workflow type. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon Simple Workflow Developer Guide.

task_start_to_close_timeout: Option<String>

Specifies the maximum duration of decision tasks for the new workflow execution. This parameter overrides the defaultTaskStartToCloseTimout specified when registering the workflow type using RegisterWorkflowType.

The duration is specified in seconds; an integer greater than or equal to 0. The value "NONE" can be used to specify unlimited duration.
A task start-to-close timeout for the new workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default task start-to-close timeout was specified at registration time then a fault will be returned.

workflow_type_version: Option<String>