Struct rusoto_swf::CountClosedWorkflowExecutionsInput

pub struct CountClosedWorkflowExecutionsInput {
    pub close_status_filter: Option<CloseStatusFilter>,
    pub close_time_filter: Option<ExecutionTimeFilter>,
    pub domain: String,
    pub execution_filter: Option<WorkflowExecutionFilter>,
    pub start_time_filter: Option<ExecutionTimeFilter>,
    pub tag_filter: Option<TagFilter>,
    pub type_filter: Option<WorkflowTypeFilter>,
}

Fields
close_status_filter: Option<CloseStatusFilter>

If specified, only workflow executions that match this close status are counted. This filter has an affect only if executionStatus is specified as CLOSED.
closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

close_time_filter: Option<ExecutionTimeFilter>

If specified, only workflow executions that meet the close time criteria of the filter are counted.
startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

domain: String

The name of the domain containing the workflow executions to count.

execution_filter: Option<WorkflowExecutionFilter>

If specified, only workflow executions matching the WorkflowId in the filter are counted.
closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

start_time_filter: Option<ExecutionTimeFilter>

If specified, only workflow executions that meet the start time criteria of the filter are counted.
startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

tag_filter: Option<TagFilter>

If specified, only executions that have a tag that matches the filter are counted.
closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

type_filter: Option<WorkflowTypeFilter>

If specified, indicates the type of the workflow executions to be counted.
closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.