Struct rusoto_swf::ListDomainsInput

pub struct ListDomainsInput {
    pub maximum_page_size: Option<i64>,
    pub next_page_token: Option<String>,
    pub registration_status: String,
    pub reverse_order: Option<bool>,
}

Fields
maximum_page_size: Option<i64>

The maximum number of results that will be returned per call. nextPageToken can be used to obtain futher pages of results. The default is 1000, which is the maximum allowed page size. You can, however, specify a page size smaller than the maximum.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

next_page_token: Option<String>

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

registration_status: String

Specifies the registration status of the domains to list.

reverse_order: Option<bool>

When set to true, returns the results in reverse order. By default, the results are returned in ascending alphabetical order by name of the domains.