Struct rusoto_swf::SwfClient

pub struct SwfClient { /* fields omitted */ }

A client for the Amazon SWF API.

Methods

impl SwfClient

pub fn new(region: Region) -> SwfClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> SwfClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,