Struct rusoto_swf::WorkflowExecutionConfiguration

pub struct WorkflowExecutionConfiguration {
    pub child_policy: String,
    pub execution_start_to_close_timeout: String,
    pub lambda_role: Option<String>,
    pub task_list: TaskList,
    pub task_priority: Option<String>,
    pub task_start_to_close_timeout: String,
}

The configuration settings for a workflow execution including timeout values, tasklist etc. These configuration settings are determined from the defaults specified when registering the workflow type and those specified when starting the workflow execution.

Fields
child_policy: String

The policy to use for the child workflow executions if this workflow execution is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout.

The supported child policies are:

TERMINATE: the child executions will be terminated.

REQUEST_CANCEL: a request to cancel will be attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON: no action will be taken. The child executions will continue to run.

execution_start_to_close_timeout: String

The total duration for this workflow execution.

The duration is specified in seconds; an integer greater than or equal to 0. The value "NONE" can be used to specify unlimited duration.

lambda_role: Option<String>

The IAM role used by this workflow execution when invoking AWS Lambda functions.

task_list: TaskList

The task list used for the decision tasks generated for this workflow execution.

task_priority: Option<String>

The priority assigned to decision tasks for this workflow execution. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon Simple Workflow Developer Guide.

task_start_to_close_timeout: String

The maximum duration allowed for decision tasks for this workflow execution.

The duration is specified in seconds; an integer greater than or equal to 0. The value "NONE" can be used to specify unlimited duration.