Struct rusoto_swf::WorkflowTypeConfiguration

pub struct WorkflowTypeConfiguration {
    pub default_child_policy: Option<String>,
    pub default_execution_start_to_close_timeout: Option<String>,
    pub default_lambda_role: Option<String>,
    pub default_task_list: Option<TaskList>,
    pub default_task_priority: Option<String>,
    pub default_task_start_to_close_timeout: Option<String>,
}

The configuration settings of a workflow type.

Fields
default_child_policy: Option<String>

Optional. The default policy to use for the child workflow executions when a workflow execution of this type is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution decision.

The supported child policies are:

TERMINATE: the child executions will be terminated.

REQUEST_CANCEL: a request to cancel will be attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON: no action will be taken. The child executions will continue to run.

default_execution_start_to_close_timeout: Option<String>

Optional. The default maximum duration, specified when registering the workflow type, for executions of this workflow type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution decision.

The duration is specified in seconds; an integer greater than or equal to 0. The value "NONE" can be used to specify unlimited duration.

default_lambda_role: Option<String>

The default IAM role to use when a workflow execution invokes a AWS Lambda function.

default_task_list: Option<TaskList>

Optional. The default task list, specified when registering the workflow type, for decisions tasks scheduled for workflow executions of this type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution decision.

default_task_priority: Option<String>

Optional. The default task priority, specified when registering the workflow type, for all decision tasks of this workflow type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution decision.

Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon Simple Workflow Developer Guide.

default_task_start_to_close_timeout: Option<String>

Optional. The default maximum duration, specified when registering the workflow type, that a decision task for executions of this workflow type might take before returning completion or failure. If the task does not close in the specified time then the task is automatically timed out and rescheduled. If the decider eventually reports a completion or failure, it is ignored. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution decision.

The duration is specified in seconds; an integer greater than or equal to 0. The value "NONE" can be used to specify unlimited duration.