Crate rusoto_swf

Amazon Simple Workflow Service
The Amazon Simple Workflow Service (Amazon SWF) makes it easy to build applications that use Amazon's cloud to coordinate work across distributed components. In Amazon SWF, a task represents a logical unit of work that is performed by a component of your workflow. Coordinating tasks in a workflow involves managing intertask dependencies, scheduling, and concurrency in accordance with the logical flow of the application.

Amazon SWF gives you full control over implementing tasks and coordinating them without worrying about underlying complexities such as tracking their progress and maintaining their state.

This documentation serves as reference only. For a broader overview of the Amazon SWF programming model, see the Amazon SWF Developer Guide.

If you're using the service, you're probably looking for SwfClient and Swf.

Structs

ActivityTask
Unit of work sent to an activity worker.

ActivityTaskCancelRequestedEventAttributes
Provides details of the ActivityTaskCancelRequested event.

ActivityTaskCanceledEventAttributes
Provides details of the ActivityTaskCanceled event.

ActivityTaskCompletedEventAttributes
Provides details of the ActivityTaskCompleted event.

ActivityTaskFailedEventAttributes
Provides details of the ActivityTaskFailed event.

ActivityTaskScheduledEventAttributes
Provides details of the ActivityTaskScheduled event.

ActivityTaskStartedEventAttributes
Provides details of the ActivityTaskStarted event.

ActivityTaskStatus
Status information about an activity task.

ActivityTaskTimedOutEventAttributes
Provides details of the ActivityTaskTimedOut event.

ActivityType
Represents an activity type.

ActivityTypeConfiguration
Configuration settings registered with the activity type.

ActivityTypeDetail
Detailed information about an activity type.

ActivityTypeInfo
Detailed information about an activity type.

ActivityTypeInfos
Contains a paginated list of activity type information structures.

CancelTimerDecisionAttributes
Provides details of the CancelTimer decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

CancelTimerFailedEventAttributes
Provides details of the CancelTimerFailed event.

CancelWorkflowExecutionDecisionAttributes
Provides details of the CancelWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

CancelWorkflowExecutionFailedEventAttributes
Provides details of the CancelWorkflowExecutionFailed event.

ChildWorkflowExecutionCanceledEventAttributes
Provide details of the ChildWorkflowExecutionCanceled event.

ChildWorkflowExecutionCompletedEventAttributes
Provides details of the ChildWorkflowExecutionCompleted event.

ChildWorkflowExecutionFailedEventAttributes
Provides details of the ChildWorkflowExecutionFailed event.

ChildWorkflowExecutionStartedEventAttributes
Provides details of the ChildWorkflowExecutionStarted event.

ChildWorkflowExecutionTerminatedEventAttributes
Provides details of the ChildWorkflowExecutionTerminated event.

ChildWorkflowExecutionTimedOutEventAttributes
Provides details of the ChildWorkflowExecutionTimedOut event.

CloseStatusFilter
Used to filter the closed workflow executions in visibility APIs by their close status.

CompleteWorkflowExecutionDecisionAttributes
Provides details of the CompleteWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

CompleteWorkflowExecutionFailedEventAttributes
Provides details of the CompleteWorkflowExecutionFailed event.

ContinueAsNewWorkflowExecutionDecisionAttributes
Provides details of the ContinueAsNewWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tag: Optional.. A tag used to identify the workflow execution

taskList: String constraint. The key is swf:taskList.name.

workflowType.version: String constraint. The key is swf:workflowType.version.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

ContinueAsNewWorkflowExecutionFailedEventAttributes
Provides details of the ContinueAsNewWorkflowExecutionFailed event.

CountClosedWorkflowExecutionsInput

CountOpenWorkflowExecutionsInput

CountPendingActivityTasksInput

CountPendingDecisionTasksInput

Decision

Specifies a decision made by the decider. A decision can be one of these types:

CancelTimer: cancels a previously started timer and records a TimerCanceled event in the history.

CancelWorkflowExecution: closes the workflow execution and records a WorkflowExecutionCanceled event in the history.

CompleteWorkflowExecution: closes the workflow execution and records a WorkflowExecutionCompleted event in the history .

ContinueAsNewWorkflowExecution: closes the workflow execution and starts a new workflow execution of the same type using the same workflow ID and a unique run ID. A WorkflowExecutionContinuedAsNew event is recorded in the history.

FailWorkflowExecution: closes the workflow execution and records a WorkflowExecutionFailed event in the history.

RecordMarker: records a MarkerRecorded event in the history. Markers can be used for adding custom information in the history for instance to let deciders know that they do not need to look at the history beyond the marker event.

RequestCancelActivityTask: attempts to cancel a previously scheduled activity task. If the activity task was scheduled but has not been assigned to a worker, then it will be canceled. If the activity task was already assigned to a worker, then the worker will be informed that cancellation has been requested in the response to RecordActivityTaskHeartbeat.

RequestCancelExternalWorkflowExecution: requests that a request be made to cancel the specified external workflow execution and records a RequestCancelExternalWorkflowExecutionInitiated event in the history.

ScheduleActivityTask: schedules an activity task.

ScheduleLambdaFunction: schedules a AWS Lambda function.

SignalExternalWorkflowExecution: requests a signal to be delivered to the specified external workflow execution and records a SignalExternalWorkflowExecutionInitiated event in the history.

StartChildWorkflowExecution: requests that a child workflow execution be started and records a StartChildWorkflowExecutionInitiated event in the history. The child workflow execution is a separate workflow execution with its own history.

StartTimer: starts a timer for this workflow execution and records a TimerStarted event in the history. This timer will fire after the specified delay and record a TimerFired event.

Access Control

If you grant permission to use RespondDecisionTaskCompleted, you can use IAM policies to express permissions for the list of decisions returned by this action as if they were members of the API. Treating decisions as a pseudo API maintains a uniform conceptual model and helps keep policies readable. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

Decision Failure

Decisions can fail for several reasons

The ordering of decisions should follow a logical flow. Some decisions might not make sense in the current context of the workflow execution and will therefore fail.

A limit on your account was reached.

The decision lacks sufficient permissions.

One of the following events might be added to the history to indicate an error. The event attribute's cause parameter indicates the cause. If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

ScheduleActivityTaskFailed: a ScheduleActivityTask decision failed. This could happen if the activity type specified in the decision is not registered, is in a deprecated state, or the decision is not properly configured.

ScheduleLambdaFunctionFailed: a ScheduleLambdaFunctionFailed decision failed. This could happen if the AWS Lambda function specified in the decision does not exist, or the AWS Lambda service's limits are exceeded.

RequestCancelActivityTaskFailed: a RequestCancelActivityTask decision failed. This could happen if there is no open activity task with the specified activityId.

StartTimerFailed: a StartTimer decision failed. This could happen if there is another open timer with the same timerId.

CancelTimerFailed: a CancelTimer decision failed. This could happen if there is no open timer with the specified timerId.

StartChildWorkflowExecutionFailed: a StartChildWorkflowExecution decision failed. This could happen if the workflow type specified is not registered, is deprecated, or the decision is not properly configured.

SignalExternalWorkflowExecutionFailed: a SignalExternalWorkflowExecution decision failed. This could happen if the workflowID specified in the decision was incorrect.

RequestCancelExternalWorkflowExecutionFailed: a RequestCancelExternalWorkflowExecution decision failed. This could happen if the workflowID specified in the decision was incorrect.

CancelWorkflowExecutionFailed: a CancelWorkflowExecution decision failed. This could happen if there is an unhandled decision task pending in the workflow execution.

CompleteWorkflowExecutionFailed: a CompleteWorkflowExecution decision failed. This could happen if there is an unhandled decision task pending in the workflow execution.

ContinueAsNewWorkflowExecutionFailed: a ContinueAsNewWorkflowExecution decision failed. This could happen if there is an unhandled decision task pending in the workflow execution or the ContinueAsNewWorkflowExecution decision was not configured correctly.

FailWorkflowExecutionFailed: a FailWorkflowExecution decision failed. This could happen if there is an unhandled decision task pending in the workflow execution.

The preceding error events might occur due to an error in the decider logic, which might put the workflow execution in an unstable state The cause field in the event structure for the error event indicates the cause of the error.
A workflow execution may be closed by the decider by returning one of the following decisions when completing a decision task: CompleteWorkflowExecution, FailWorkflowExecution, CancelWorkflowExecution and ContinueAsNewWorkflowExecution. An UnhandledDecision fault will be returned if a workflow closing decision is specified and a signal or activity event had been added to the history while the decision task was being performed by the decider. Unlike the above situations which are logic issues, this fault is always possible because of race conditions in a distributed system. The right action here is to call RespondDecisionTaskCompleted without any decisions. This would result in another decision task with these new events included in the history. The decider should handle the new events and may decide to close the workflow execution.
How to code a decision

You code a decision by first setting the decision type field to one of the above decision values, and then set the corresponding attributes field shown below:

ScheduleActivityTaskDecisionAttributes

ScheduleLambdaFunctionDecisionAttributes

RequestCancelActivityTaskDecisionAttributes

CompleteWorkflowExecutionDecisionAttributes

FailWorkflowExecutionDecisionAttributes

CancelWorkflowExecutionDecisionAttributes

ContinueAsNewWorkflowExecutionDecisionAttributes

RecordMarkerDecisionAttributes

StartTimerDecisionAttributes

CancelTimerDecisionAttributes

SignalExternalWorkflowExecutionDecisionAttributes

RequestCancelExternalWorkflowExecutionDecisionAttributes

StartChildWorkflowExecutionDecisionAttributes

DecisionTask
A structure that represents a decision task. Decision tasks are sent to deciders in order for them to make decisions.

DecisionTaskCompletedEventAttributes
Provides details of the DecisionTaskCompleted event.

DecisionTaskScheduledEventAttributes
Provides details about the DecisionTaskScheduled event.

DecisionTaskStartedEventAttributes
Provides details of the DecisionTaskStarted event.

DecisionTaskTimedOutEventAttributes
Provides details of the DecisionTaskTimedOut event.

DeprecateActivityTypeInput

DeprecateDomainInput

DeprecateWorkflowTypeInput

DescribeActivityTypeInput

DescribeDomainInput

DescribeWorkflowExecutionInput

DescribeWorkflowTypeInput

DomainConfiguration
Contains the configuration settings of a domain.

DomainDetail
Contains details of a domain.

DomainInfo
Contains general information about a domain.

DomainInfos
Contains a paginated collection of DomainInfo structures.

ExecutionTimeFilter
Used to filter the workflow executions in visibility APIs by various time-based rules. Each parameter, if specified, defines a rule that must be satisfied by each returned query result. The parameter values are in the Unix Time format. For example: "oldestDate": 1325376070.

ExternalWorkflowExecutionCancelRequestedEventAttributes
Provides details of the ExternalWorkflowExecutionCancelRequested event.

ExternalWorkflowExecutionSignaledEventAttributes
Provides details of the ExternalWorkflowExecutionSignaled event.

FailWorkflowExecutionDecisionAttributes
Provides details of the FailWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

FailWorkflowExecutionFailedEventAttributes
Provides details of the FailWorkflowExecutionFailed event.

GetWorkflowExecutionHistoryInput

History
Paginated representation of a workflow history for a workflow execution. This is the up to date, complete and authoritative record of the events related to all tasks and events in the life of the workflow execution.

HistoryEvent

Event within a workflow execution. A history event can be one of these types:

WorkflowExecutionStarted: The workflow execution was started.

WorkflowExecutionCompleted: The workflow execution was closed due to successful completion.

WorkflowExecutionFailed: The workflow execution closed due to a failure.

WorkflowExecutionTimedOut: The workflow execution was closed because a time out was exceeded.

WorkflowExecutionCanceled: The workflow execution was successfully canceled and closed.

WorkflowExecutionTerminated: The workflow execution was terminated.

WorkflowExecutionContinuedAsNew: The workflow execution was closed and a new execution of the same type was created with the same workflowId.

WorkflowExecutionCancelRequested: A request to cancel this workflow execution was made.

DecisionTaskScheduled: A decision task was scheduled for the workflow execution.

DecisionTaskStarted: The decision task was dispatched to a decider.

DecisionTaskCompleted: The decider successfully completed a decision task by calling RespondDecisionTaskCompleted.

DecisionTaskTimedOut: The decision task timed out.

ActivityTaskScheduled: An activity task was scheduled for execution.

ScheduleActivityTaskFailed: Failed to process ScheduleActivityTask decision. This happens when the decision is not configured properly, for example the activity type specified is not registered.

ActivityTaskStarted: The scheduled activity task was dispatched to a worker.

ActivityTaskCompleted: An activity worker successfully completed an activity task by calling RespondActivityTaskCompleted.

ActivityTaskFailed: An activity worker failed an activity task by calling RespondActivityTaskFailed.

ActivityTaskTimedOut: The activity task timed out.

ActivityTaskCanceled: The activity task was successfully canceled.

ActivityTaskCancelRequested: A RequestCancelActivityTask decision was received by the system.

RequestCancelActivityTaskFailed: Failed to process RequestCancelActivityTask decision. This happens when the decision is not configured properly.

WorkflowExecutionSignaled: An external signal was received for the workflow execution.

MarkerRecorded: A marker was recorded in the workflow history as the result of a RecordMarker decision.

TimerStarted: A timer was started for the workflow execution due to a StartTimer decision.

StartTimerFailed: Failed to process StartTimer decision. This happens when the decision is not configured properly, for example a timer already exists with the specified timer ID.

TimerFired: A timer, previously started for this workflow execution, fired.

TimerCanceled: A timer, previously started for this workflow execution, was successfully canceled.

CancelTimerFailed: Failed to process CancelTimer decision. This happens when the decision is not configured properly, for example no timer exists with the specified timer ID.

StartChildWorkflowExecutionInitiated: A request was made to start a child workflow execution.

StartChildWorkflowExecutionFailed: Failed to process StartChildWorkflowExecution decision. This happens when the decision is not configured properly, for example the workflow type specified is not registered.

ChildWorkflowExecutionStarted: A child workflow execution was successfully started.

ChildWorkflowExecutionCompleted: A child workflow execution, started by this workflow execution, completed successfully and was closed.

ChildWorkflowExecutionFailed: A child workflow execution, started by this workflow execution, failed to complete successfully and was closed.

ChildWorkflowExecutionTimedOut: A child workflow execution, started by this workflow execution, timed out and was closed.

ChildWorkflowExecutionCanceled: A child workflow execution, started by this workflow execution, was canceled and closed.

ChildWorkflowExecutionTerminated: A child workflow execution, started by this workflow execution, was terminated.

SignalExternalWorkflowExecutionInitiated: A request to signal an external workflow was made.

ExternalWorkflowExecutionSignaled: A signal, requested by this workflow execution, was successfully delivered to the target external workflow execution.

SignalExternalWorkflowExecutionFailed: The request to signal an external workflow execution failed.

RequestCancelExternalWorkflowExecutionInitiated: A request was made to request the cancellation of an external workflow execution.

ExternalWorkflowExecutionCancelRequested: Request to cancel an external workflow execution was successfully delivered to the target execution.

RequestCancelExternalWorkflowExecutionFailed: Request to cancel an external workflow execution failed.

LambdaFunctionScheduled: An AWS Lambda function was scheduled for execution.

LambdaFunctionStarted: The scheduled function was invoked in the AWS Lambda service.

LambdaFunctionCompleted: The AWS Lambda function successfully completed.

LambdaFunctionFailed: The AWS Lambda function execution failed.

LambdaFunctionTimedOut: The AWS Lambda function execution timed out.

ScheduleLambdaFunctionFailed: Failed to process ScheduleLambdaFunction decision. This happens when the workflow execution does not have the proper IAM role attached to invoke AWS Lambda functions.

StartLambdaFunctionFailed: Failed to invoke the scheduled function in the AWS Lambda service. This happens when the AWS Lambda service is not available in the current region, or received too many requests.

LambdaFunctionCompletedEventAttributes
Provides details for the LambdaFunctionCompleted event.

LambdaFunctionFailedEventAttributes
Provides details for the LambdaFunctionFailed event.

LambdaFunctionScheduledEventAttributes
Provides details for the LambdaFunctionScheduled event.

LambdaFunctionStartedEventAttributes
Provides details for the LambdaFunctionStarted event.

LambdaFunctionTimedOutEventAttributes
Provides details for the LambdaFunctionTimedOut event.

ListActivityTypesInput

ListClosedWorkflowExecutionsInput

ListDomainsInput

ListOpenWorkflowExecutionsInput

ListWorkflowTypesInput

MarkerRecordedEventAttributes
Provides details of the MarkerRecorded event.

PendingTaskCount
Contains the count of tasks in a task list.

PollForActivityTaskInput

PollForDecisionTaskInput

RecordActivityTaskHeartbeatInput

RecordMarkerDecisionAttributes
Provides details of the RecordMarker decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

RecordMarkerFailedEventAttributes
Provides details of the RecordMarkerFailed event.

RegisterActivityTypeInput

RegisterDomainInput

RegisterWorkflowTypeInput

RequestCancelActivityTaskDecisionAttributes
Provides details of the RequestCancelActivityTask decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

RequestCancelActivityTaskFailedEventAttributes
Provides details of the RequestCancelActivityTaskFailed event.

RequestCancelExternalWorkflowExecutionDecisionAttributes
Provides details of the RequestCancelExternalWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

RequestCancelExternalWorkflowExecutionFailedEventAttributes
Provides details of the RequestCancelExternalWorkflowExecutionFailed event.

RequestCancelExternalWorkflowExecutionInitiatedEventAttributes
Provides details of the RequestCancelExternalWorkflowExecutionInitiated event.

RequestCancelWorkflowExecutionInput

RespondActivityTaskCanceledInput

RespondActivityTaskCompletedInput

RespondActivityTaskFailedInput

RespondDecisionTaskCompletedInput

Run
Specifies the runId of a workflow execution.

ScheduleActivityTaskDecisionAttributes
Provides details of the ScheduleActivityTask decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

activityType.name: String constraint. The key is swf:activityType.name.

activityType.version: String constraint. The key is swf:activityType.version.

taskList: String constraint. The key is swf:taskList.name.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

ScheduleActivityTaskFailedEventAttributes
Provides details of the ScheduleActivityTaskFailed event.

ScheduleLambdaFunctionDecisionAttributes
Provides details of the ScheduleLambdaFunction decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

activityType.name: String constraint. The key is swf:activityType.name.

activityType.version: String constraint. The key is swf:activityType.version.

taskList: String constraint. The key is swf:taskList.name.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

ScheduleLambdaFunctionFailedEventAttributes
Provides details for the ScheduleLambdaFunctionFailed event.

SignalExternalWorkflowExecutionDecisionAttributes
Provides details of the SignalExternalWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

SignalExternalWorkflowExecutionFailedEventAttributes
Provides details of the SignalExternalWorkflowExecutionFailed event.

SignalExternalWorkflowExecutionInitiatedEventAttributes
Provides details of the SignalExternalWorkflowExecutionInitiated event.

SignalWorkflowExecutionInput

StartChildWorkflowExecutionDecisionAttributes
Provides details of the StartChildWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tagList.member.N: The key is "swf:tagList.N" where N is the tag number from 0 to 4, inclusive.

taskList: String constraint. The key is swf:taskList.name.

workflowType.name: String constraint. The key is swf:workflowType.name.

workflowType.version: String constraint. The key is swf:workflowType.version.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

StartChildWorkflowExecutionFailedEventAttributes
Provides details of the StartChildWorkflowExecutionFailed event.

StartChildWorkflowExecutionInitiatedEventAttributes
Provides details of the StartChildWorkflowExecutionInitiated event.

StartLambdaFunctionFailedEventAttributes
Provides details for the StartLambdaFunctionFailed event.

StartTimerDecisionAttributes
Provides details of the StartTimer decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller does not have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter will be set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows.

StartTimerFailedEventAttributes
Provides details of the StartTimerFailed event.

StartWorkflowExecutionInput

SwfClient
A client for the Amazon SWF API.

TagFilter
Used to filter the workflow executions in visibility APIs based on a tag.

TaskList
Represents a task list.

TerminateWorkflowExecutionInput

TimerCanceledEventAttributes
Provides details of the TimerCanceled event.

TimerFiredEventAttributes
Provides details of the TimerFired event.

TimerStartedEventAttributes
Provides details of the TimerStarted event.

WorkflowExecution
Represents a workflow execution.

WorkflowExecutionCancelRequestedEventAttributes
Provides details of the WorkflowExecutionCancelRequested event.

WorkflowExecutionCanceledEventAttributes
Provides details of the WorkflowExecutionCanceled event.

WorkflowExecutionCompletedEventAttributes
Provides details of the WorkflowExecutionCompleted event.

WorkflowExecutionConfiguration
The configuration settings for a workflow execution including timeout values, tasklist etc. These configuration settings are determined from the defaults specified when registering the workflow type and those specified when starting the workflow execution.

WorkflowExecutionContinuedAsNewEventAttributes
Provides details of the WorkflowExecutionContinuedAsNew event.

WorkflowExecutionCount
Contains the count of workflow executions returned from CountOpenWorkflowExecutions or CountClosedWorkflowExecutions

WorkflowExecutionDetail
Contains details about a workflow execution.

WorkflowExecutionFailedEventAttributes
Provides details of the WorkflowExecutionFailed event.

WorkflowExecutionFilter
Used to filter the workflow executions in visibility APIs by their workflowId.

WorkflowExecutionInfo
Contains information about a workflow execution.

WorkflowExecutionInfos
Contains a paginated list of information about workflow executions.

WorkflowExecutionOpenCounts
Contains the counts of open tasks, child workflow executions and timers for a workflow execution.

WorkflowExecutionSignaledEventAttributes
Provides details of the WorkflowExecutionSignaled event.

WorkflowExecutionStartedEventAttributes
Provides details of WorkflowExecutionStarted event.

WorkflowExecutionTerminatedEventAttributes
Provides details of the WorkflowExecutionTerminated event.

WorkflowExecutionTimedOutEventAttributes
Provides details of the WorkflowExecutionTimedOut event.

WorkflowType
Represents a workflow type.

WorkflowTypeConfiguration
The configuration settings of a workflow type.

WorkflowTypeDetail
Contains details about a workflow type.

WorkflowTypeFilter
Used to filter workflow execution query results by type. Each parameter, if specified, defines a rule that must be satisfied by each returned result.

WorkflowTypeInfo
Contains information about a workflow type.

WorkflowTypeInfos
Contains a paginated list of information structures about workflow types.

Enums

CountClosedWorkflowExecutionsError
Errors returned by CountClosedWorkflowExecutions

CountOpenWorkflowExecutionsError
Errors returned by CountOpenWorkflowExecutions

CountPendingActivityTasksError
Errors returned by CountPendingActivityTasks

CountPendingDecisionTasksError
Errors returned by CountPendingDecisionTasks

DeprecateActivityTypeError
Errors returned by DeprecateActivityType

DeprecateDomainError
Errors returned by DeprecateDomain

DeprecateWorkflowTypeError
Errors returned by DeprecateWorkflowType

DescribeActivityTypeError
Errors returned by DescribeActivityType

DescribeDomainError
Errors returned by DescribeDomain

DescribeWorkflowExecutionError
Errors returned by DescribeWorkflowExecution

DescribeWorkflowTypeError
Errors returned by DescribeWorkflowType

GetWorkflowExecutionHistoryError
Errors returned by GetWorkflowExecutionHistory

ListActivityTypesError
Errors returned by ListActivityTypes

ListClosedWorkflowExecutionsError
Errors returned by ListClosedWorkflowExecutions

ListDomainsError
Errors returned by ListDomains

ListOpenWorkflowExecutionsError
Errors returned by ListOpenWorkflowExecutions

ListWorkflowTypesError
Errors returned by ListWorkflowTypes

PollForActivityTaskError
Errors returned by PollForActivityTask

PollForDecisionTaskError
Errors returned by PollForDecisionTask

RecordActivityTaskHeartbeatError
Errors returned by RecordActivityTaskHeartbeat

RegisterActivityTypeError
Errors returned by RegisterActivityType

RegisterDomainError
Errors returned by RegisterDomain

RegisterWorkflowTypeError
Errors returned by RegisterWorkflowType

RequestCancelWorkflowExecutionError
Errors returned by RequestCancelWorkflowExecution

RespondActivityTaskCanceledError
Errors returned by RespondActivityTaskCanceled

RespondActivityTaskCompletedError
Errors returned by RespondActivityTaskCompleted

RespondActivityTaskFailedError
Errors returned by RespondActivityTaskFailed

RespondDecisionTaskCompletedError
Errors returned by RespondDecisionTaskCompleted

SignalWorkflowExecutionError
Errors returned by SignalWorkflowExecution

StartWorkflowExecutionError
Errors returned by StartWorkflowExecution

TerminateWorkflowExecutionError
Errors returned by TerminateWorkflowExecution

Traits

Swf
Trait representing the capabilities of the Amazon SWF API. Amazon SWF clients implement this trait.