Struct rusoto_swf::HistoryEvent

pub struct HistoryEvent {
    pub activity_task_cancel_requested_event_attributes: Option<ActivityTaskCancelRequestedEventAttributes>,
    pub activity_task_canceled_event_attributes: Option<ActivityTaskCanceledEventAttributes>,
    pub activity_task_completed_event_attributes: Option<ActivityTaskCompletedEventAttributes>,
    pub activity_task_failed_event_attributes: Option<ActivityTaskFailedEventAttributes>,
    pub activity_task_scheduled_event_attributes: Option<ActivityTaskScheduledEventAttributes>,
    pub activity_task_started_event_attributes: Option<ActivityTaskStartedEventAttributes>,
    pub activity_task_timed_out_event_attributes: Option<ActivityTaskTimedOutEventAttributes>,
    pub cancel_timer_failed_event_attributes: Option<CancelTimerFailedEventAttributes>,
    pub cancel_workflow_execution_failed_event_attributes: Option<CancelWorkflowExecutionFailedEventAttributes>,
    pub child_workflow_execution_canceled_event_attributes: Option<ChildWorkflowExecutionCanceledEventAttributes>,
    pub child_workflow_execution_completed_event_attributes: Option<ChildWorkflowExecutionCompletedEventAttributes>,
    pub child_workflow_execution_failed_event_attributes: Option<ChildWorkflowExecutionFailedEventAttributes>,
    pub child_workflow_execution_started_event_attributes: Option<ChildWorkflowExecutionStartedEventAttributes>,
    pub child_workflow_execution_terminated_event_attributes: Option<ChildWorkflowExecutionTerminatedEventAttributes>,
    pub child_workflow_execution_timed_out_event_attributes: Option<ChildWorkflowExecutionTimedOutEventAttributes>,
    pub complete_workflow_execution_failed_event_attributes: Option<CompleteWorkflowExecutionFailedEventAttributes>,
    pub continue_as_new_workflow_execution_failed_event_attributes: Option<ContinueAsNewWorkflowExecutionFailedEventAttributes>,
    pub decision_task_completed_event_attributes: Option<DecisionTaskCompletedEventAttributes>,
    pub decision_task_scheduled_event_attributes: Option<DecisionTaskScheduledEventAttributes>,
    pub decision_task_started_event_attributes: Option<DecisionTaskStartedEventAttributes>,
    pub decision_task_timed_out_event_attributes: Option<DecisionTaskTimedOutEventAttributes>,
    pub event_id: i64,
    pub event_timestamp: f64,
    pub event_type: String,
    pub external_workflow_execution_cancel_requested_event_attributes: Option<ExternalWorkflowExecutionCancelRequestedEventAttributes>,
    pub external_workflow_execution_signaled_event_attributes: Option<ExternalWorkflowExecutionSignaledEventAttributes>,
    pub fail_workflow_execution_failed_event_attributes: Option<FailWorkflowExecutionFailedEventAttributes>,
    pub lambda_function_completed_event_attributes: Option<LambdaFunctionCompletedEventAttributes>,
    pub lambda_function_failed_event_attributes: Option<LambdaFunctionFailedEventAttributes>,
    pub lambda_function_scheduled_event_attributes: Option<LambdaFunctionScheduledEventAttributes>,
    pub lambda_function_started_event_attributes: Option<LambdaFunctionStartedEventAttributes>,
    pub lambda_function_timed_out_event_attributes: Option<LambdaFunctionTimedOutEventAttributes>,
    pub marker_recorded_event_attributes: Option<MarkerRecordedEventAttributes>,
    pub record_marker_failed_event_attributes: Option<RecordMarkerFailedEventAttributes>,
    pub request_cancel_activity_task_failed_event_attributes: Option<RequestCancelActivityTaskFailedEventAttributes>,
    pub request_cancel_external_workflow_execution_failed_event_attributes: Option<RequestCancelExternalWorkflowExecutionFailedEventAttributes>,
    pub request_cancel_external_workflow_execution_initiated_event_attributes: Option<RequestCancelExternalWorkflowExecutionInitiatedEventAttributes>,
    pub schedule_activity_task_failed_event_attributes: Option<ScheduleActivityTaskFailedEventAttributes>,
    pub schedule_lambda_function_failed_event_attributes: Option<ScheduleLambdaFunctionFailedEventAttributes>,
    pub signal_external_workflow_execution_failed_event_attributes: Option<SignalExternalWorkflowExecutionFailedEventAttributes>,
    pub signal_external_workflow_execution_initiated_event_attributes: Option<SignalExternalWorkflowExecutionInitiatedEventAttributes>,
    pub start_child_workflow_execution_failed_event_attributes: Option<StartChildWorkflowExecutionFailedEventAttributes>,
    pub start_child_workflow_execution_initiated_event_attributes: Option<StartChildWorkflowExecutionInitiatedEventAttributes>,
    pub start_lambda_function_failed_event_attributes: Option<StartLambdaFunctionFailedEventAttributes>,
    pub start_timer_failed_event_attributes: Option<StartTimerFailedEventAttributes>,
    pub timer_canceled_event_attributes: Option<TimerCanceledEventAttributes>,
    pub timer_fired_event_attributes: Option<TimerFiredEventAttributes>,
    pub timer_started_event_attributes: Option<TimerStartedEventAttributes>,
    pub workflow_execution_cancel_requested_event_attributes: Option<WorkflowExecutionCancelRequestedEventAttributes>,
    pub workflow_execution_canceled_event_attributes: Option<WorkflowExecutionCanceledEventAttributes>,
    pub workflow_execution_completed_event_attributes: Option<WorkflowExecutionCompletedEventAttributes>,
    pub workflow_execution_continued_as_new_event_attributes: Option<WorkflowExecutionContinuedAsNewEventAttributes>,
    pub workflow_execution_failed_event_attributes: Option<WorkflowExecutionFailedEventAttributes>,
    pub workflow_execution_signaled_event_attributes: Option<WorkflowExecutionSignaledEventAttributes>,
    pub workflow_execution_started_event_attributes: Option<WorkflowExecutionStartedEventAttributes>,
    pub workflow_execution_terminated_event_attributes: Option<WorkflowExecutionTerminatedEventAttributes>,
    pub workflow_execution_timed_out_event_attributes: Option<WorkflowExecutionTimedOutEventAttributes>,
}

Event within a workflow execution. A history event can be one of these types:

ActivityTaskCancelRequested – A RequestCancelActivityTask decision was received by the system.

ActivityTaskCanceled – The activity task was successfully canceled.

ActivityTaskCompleted – An activity worker successfully completed an activity task by calling RespondActivityTaskCompleted.

ActivityTaskFailed – An activity worker failed an activity task by calling RespondActivityTaskFailed.

ActivityTaskScheduled – An activity task was scheduled for execution.

ActivityTaskStarted – The scheduled activity task was dispatched to a worker.

ActivityTaskTimedOut – The activity task timed out.

CancelTimerFailed – Failed to process CancelTimer decision. This happens when the decision isn't configured properly, for example no timer exists with the specified timer Id.

CancelWorkflowExecutionFailed – A request to cancel a workflow execution failed.

ChildWorkflowExecutionCanceled – A child workflow execution, started by this workflow execution, was canceled and closed.

ChildWorkflowExecutionCompleted – A child workflow execution, started by this workflow execution, completed successfully and was closed.

ChildWorkflowExecutionFailed – A child workflow execution, started by this workflow execution, failed to complete successfully and was closed.

ChildWorkflowExecutionStarted – A child workflow execution was successfully started.

ChildWorkflowExecutionTerminated – A child workflow execution, started by this workflow execution, was terminated.

ChildWorkflowExecutionTimedOut – A child workflow execution, started by this workflow execution, timed out and was closed.

CompleteWorkflowExecutionFailed – The workflow execution failed to complete.

ContinueAsNewWorkflowExecutionFailed – The workflow execution failed to complete after being continued as a new workflow execution.

DecisionTaskCompleted – The decider successfully completed a decision task by calling RespondDecisionTaskCompleted.

DecisionTaskScheduled – A decision task was scheduled for the workflow execution.

DecisionTaskStarted – The decision task was dispatched to a decider.

DecisionTaskTimedOut – The decision task timed out.

ExternalWorkflowExecutionCancelRequested – Request to cancel an external workflow execution was successfully delivered to the target execution.

ExternalWorkflowExecutionSignaled – A signal, requested by this workflow execution, was successfully delivered to the target external workflow execution.

FailWorkflowExecutionFailed – A request to mark a workflow execution as failed, itself failed.

MarkerRecorded – A marker was recorded in the workflow history as the result of a RecordMarker decision.

RecordMarkerFailed – A RecordMarker decision was returned as failed.

RequestCancelActivityTaskFailed – Failed to process RequestCancelActivityTask decision. This happens when the decision isn't configured properly.

RequestCancelExternalWorkflowExecutionFailed – Request to cancel an external workflow execution failed.

RequestCancelExternalWorkflowExecutionInitiated – A request was made to request the cancellation of an external workflow execution.

ScheduleActivityTaskFailed – Failed to process ScheduleActivityTask decision. This happens when the decision isn't configured properly, for example the activity type specified isn't registered.

SignalExternalWorkflowExecutionFailed – The request to signal an external workflow execution failed.

SignalExternalWorkflowExecutionInitiated – A request to signal an external workflow was made.

StartActivityTaskFailed – A scheduled activity task failed to start.

StartChildWorkflowExecutionFailed – Failed to process StartChildWorkflowExecution decision. This happens when the decision isn't configured properly, for example the workflow type specified isn't registered.

StartChildWorkflowExecutionInitiated – A request was made to start a child workflow execution.

StartTimerFailed – Failed to process StartTimer decision. This happens when the decision isn't configured properly, for example a timer already exists with the specified timer Id.

TimerCanceled – A timer, previously started for this workflow execution, was successfully canceled.

TimerFired – A timer, previously started for this workflow execution, fired.

TimerStarted – A timer was started for the workflow execution due to a StartTimer decision.

WorkflowExecutionCancelRequested – A request to cancel this workflow execution was made.

WorkflowExecutionCanceled – The workflow execution was successfully canceled and closed.

WorkflowExecutionCompleted – The workflow execution was closed due to successful completion.

WorkflowExecutionContinuedAsNew – The workflow execution was closed and a new execution of the same type was created with the same workflowId.

WorkflowExecutionFailed – The workflow execution closed due to a failure.

WorkflowExecutionSignaled – An external signal was received for the workflow execution.

WorkflowExecutionStarted – The workflow execution was started.

WorkflowExecutionTerminated – The workflow execution was terminated.

WorkflowExecutionTimedOut – The workflow execution was closed because a time out was exceeded.

Fields
activity_task_cancel_requested_event_attributes: Option<ActivityTaskCancelRequestedEventAttributes>

If the event is of type ActivityTaskcancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

activity_task_canceled_event_attributes: Option<ActivityTaskCanceledEventAttributes>

If the event is of type ActivityTaskCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

activity_task_completed_event_attributes: Option<ActivityTaskCompletedEventAttributes>

If the event is of type ActivityTaskCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

activity_task_failed_event_attributes: Option<ActivityTaskFailedEventAttributes>

If the event is of type ActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

activity_task_scheduled_event_attributes: Option<ActivityTaskScheduledEventAttributes>

If the event is of type ActivityTaskScheduled then this member is set and provides detailed information about the event. It isn't set for other event types.

activity_task_started_event_attributes: Option<ActivityTaskStartedEventAttributes>

If the event is of type ActivityTaskStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

activity_task_timed_out_event_attributes: Option<ActivityTaskTimedOutEventAttributes>

If the event is of type ActivityTaskTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

cancel_timer_failed_event_attributes: Option<CancelTimerFailedEventAttributes>

If the event is of type CancelTimerFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

cancel_workflow_execution_failed_event_attributes: Option<CancelWorkflowExecutionFailedEventAttributes>

If the event is of type CancelWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

child_workflow_execution_canceled_event_attributes: Option<ChildWorkflowExecutionCanceledEventAttributes>

If the event is of type ChildWorkflowExecutionCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

child_workflow_execution_completed_event_attributes: Option<ChildWorkflowExecutionCompletedEventAttributes>

If the event is of type ChildWorkflowExecutionCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

child_workflow_execution_failed_event_attributes: Option<ChildWorkflowExecutionFailedEventAttributes>

If the event is of type ChildWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

child_workflow_execution_started_event_attributes: Option<ChildWorkflowExecutionStartedEventAttributes>

If the event is of type ChildWorkflowExecutionStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

child_workflow_execution_terminated_event_attributes: Option<ChildWorkflowExecutionTerminatedEventAttributes>

If the event is of type ChildWorkflowExecutionTerminated then this member is set and provides detailed information about the event. It isn't set for other event types.

child_workflow_execution_timed_out_event_attributes: Option<ChildWorkflowExecutionTimedOutEventAttributes>

If the event is of type ChildWorkflowExecutionTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

complete_workflow_execution_failed_event_attributes: Option<CompleteWorkflowExecutionFailedEventAttributes>

If the event is of type CompleteWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

continue_as_new_workflow_execution_failed_event_attributes: Option<ContinueAsNewWorkflowExecutionFailedEventAttributes>

If the event is of type ContinueAsNewWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

decision_task_completed_event_attributes: Option<DecisionTaskCompletedEventAttributes>

If the event is of type DecisionTaskCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

decision_task_scheduled_event_attributes: Option<DecisionTaskScheduledEventAttributes>

If the event is of type DecisionTaskScheduled then this member is set and provides detailed information about the event. It isn't set for other event types.

decision_task_started_event_attributes: Option<DecisionTaskStartedEventAttributes>

If the event is of type DecisionTaskStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

decision_task_timed_out_event_attributes: Option<DecisionTaskTimedOutEventAttributes>

If the event is of type DecisionTaskTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

event_id: i64

The system generated ID of the event. This ID uniquely identifies the event with in the workflow execution history.

event_timestamp: f64

The date and time when the event occurred.

event_type: String

The type of the history event.

external_workflow_execution_cancel_requested_event_attributes: Option<ExternalWorkflowExecutionCancelRequestedEventAttributes>

If the event is of type ExternalWorkflowExecutionCancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

external_workflow_execution_signaled_event_attributes: Option<ExternalWorkflowExecutionSignaledEventAttributes>

If the event is of type ExternalWorkflowExecutionSignaled then this member is set and provides detailed information about the event. It isn't set for other event types.

fail_workflow_execution_failed_event_attributes: Option<FailWorkflowExecutionFailedEventAttributes>

If the event is of type FailWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

lambda_function_completed_event_attributes: Option<LambdaFunctionCompletedEventAttributes>

Provides the details of the LambdaFunctionCompleted event. It isn't set for other event types.

lambda_function_failed_event_attributes: Option<LambdaFunctionFailedEventAttributes>

Provides the details of the LambdaFunctionFailed event. It isn't set for other event types.

lambda_function_scheduled_event_attributes: Option<LambdaFunctionScheduledEventAttributes>

Provides the details of the LambdaFunctionScheduled event. It isn't set for other event types.

lambda_function_started_event_attributes: Option<LambdaFunctionStartedEventAttributes>

Provides the details of the LambdaFunctionStarted event. It isn't set for other event types.

lambda_function_timed_out_event_attributes: Option<LambdaFunctionTimedOutEventAttributes>

Provides the details of the LambdaFunctionTimedOut event. It isn't set for other event types.

marker_recorded_event_attributes: Option<MarkerRecordedEventAttributes>

If the event is of type MarkerRecorded then this member is set and provides detailed information about the event. It isn't set for other event types.

record_marker_failed_event_attributes: Option<RecordMarkerFailedEventAttributes>

If the event is of type DecisionTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

request_cancel_activity_task_failed_event_attributes: Option<RequestCancelActivityTaskFailedEventAttributes>

If the event is of type RequestCancelActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

request_cancel_external_workflow_execution_failed_event_attributes: Option<RequestCancelExternalWorkflowExecutionFailedEventAttributes>

If the event is of type RequestCancelExternalWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

request_cancel_external_workflow_execution_initiated_event_attributes: Option<RequestCancelExternalWorkflowExecutionInitiatedEventAttributes>

If the event is of type RequestCancelExternalWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

schedule_activity_task_failed_event_attributes: Option<ScheduleActivityTaskFailedEventAttributes>

If the event is of type ScheduleActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

schedule_lambda_function_failed_event_attributes: Option<ScheduleLambdaFunctionFailedEventAttributes>

Provides the details of the ScheduleLambdaFunctionFailed event. It isn't set for other event types.

signal_external_workflow_execution_failed_event_attributes: Option<SignalExternalWorkflowExecutionFailedEventAttributes>

If the event is of type SignalExternalWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

signal_external_workflow_execution_initiated_event_attributes: Option<SignalExternalWorkflowExecutionInitiatedEventAttributes>

If the event is of type SignalExternalWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

start_child_workflow_execution_failed_event_attributes: Option<StartChildWorkflowExecutionFailedEventAttributes>

If the event is of type StartChildWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

start_child_workflow_execution_initiated_event_attributes: Option<StartChildWorkflowExecutionInitiatedEventAttributes>

If the event is of type StartChildWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

start_lambda_function_failed_event_attributes: Option<StartLambdaFunctionFailedEventAttributes>

Provides the details of the StartLambdaFunctionFailed event. It isn't set for other event types.

start_timer_failed_event_attributes: Option<StartTimerFailedEventAttributes>

If the event is of type StartTimerFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

timer_canceled_event_attributes: Option<TimerCanceledEventAttributes>

If the event is of type TimerCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

timer_fired_event_attributes: Option<TimerFiredEventAttributes>

If the event is of type TimerFired then this member is set and provides detailed information about the event. It isn't set for other event types.

timer_started_event_attributes: Option<TimerStartedEventAttributes>

If the event is of type TimerStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

workflow_execution_cancel_requested_event_attributes: Option<WorkflowExecutionCancelRequestedEventAttributes>

If the event is of type WorkflowExecutionCancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

workflow_execution_canceled_event_attributes: Option<WorkflowExecutionCanceledEventAttributes>

If the event is of type WorkflowExecutionCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

workflow_execution_completed_event_attributes: Option<WorkflowExecutionCompletedEventAttributes>

If the event is of type WorkflowExecutionCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

workflow_execution_continued_as_new_event_attributes: Option<WorkflowExecutionContinuedAsNewEventAttributes>

If the event is of type WorkflowExecutionContinuedAsNew then this member is set and provides detailed information about the event. It isn't set for other event types.

workflow_execution_failed_event_attributes: Option<WorkflowExecutionFailedEventAttributes>

If the event is of type WorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

workflow_execution_signaled_event_attributes: Option<WorkflowExecutionSignaledEventAttributes>

If the event is of type WorkflowExecutionSignaled then this member is set and provides detailed information about the event. It isn't set for other event types.

workflow_execution_started_event_attributes: Option<WorkflowExecutionStartedEventAttributes>

If the event is of type WorkflowExecutionStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

workflow_execution_terminated_event_attributes: Option<WorkflowExecutionTerminatedEventAttributes>

If the event is of type WorkflowExecutionTerminated then this member is set and provides detailed information about the event. It isn't set for other event types.

workflow_execution_timed_out_event_attributes: Option<WorkflowExecutionTimedOutEventAttributes>

If the event is of type WorkflowExecutionTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.